feat: Tanstack start (#6)
So, wanted to do this with just HTML/CSS/JS, but man, working on it, I
really just miss components.

So, I went to SolidJS / SPA. That worked, but I'm a little hooked on SSG
now.

So I tried solid-start. Was actually happier with how they do things
there, but wasn't a fan of ditching `vite.config.ts`.

SO tanstack start, for now. solid-start are ditching devnxi or whatever
it's called and going to a standard vite route too.